+++ src/main/resources/db/migration/V001__create_table_user.sql
@@ -1,5 +1,5 @@
 create table [user](
-    id int identity primary key,
+    id bigint identity primary key,
     username nvarchar(64) not null,
     password nvarchar(64)
 );
